Updated components

+++ drivers/rtc/rtc-88pm860x.c
@@ -21,40 +21,41 @@
 
 #define VRTC_CALIBRATION
 
-struct pm860x_rtc_info {
-	struct pm860x_chip	*chip;
-	struct i2c_client	*i2c;
-	struct rtc_device	*rtc_dev;
-	struct device		*dev;
-	struct delayed_work	calib_work;
+struct pm860x_rtc_info
+{
+	struct pm860x_chip *chip;
+	struct i2c_client *i2c;
+	struct rtc_device *rtc_dev;
+	struct device *dev;
+	struct delayed_work calib_work;
 
-	int			irq;
-	int			vrtc;
-	int			(*sync)(unsigned int ticks);
+	int irq;
+	int vrtc;
+	int (*sync)(unsigned int ticks);
 };
 
-#define REG_VRTC_MEAS1		0x7D
+#define REG_VRTC_MEAS1 0x7D
 
-#define REG0_ADDR		0xB0
-#define REG1_ADDR		0xB2
-#define REG2_ADDR		0xB4
-#define REG3_ADDR		0xB6
+#define REG0_ADDR 0xB0
+#define REG1_ADDR 0xB2
+#define REG2_ADDR 0xB4
+#define REG3_ADDR 0xB6
 
-#define REG0_DATA		0xB1
-#define REG1_DATA		0xB3
-#define REG2_DATA		0xB5
-#define REG3_DATA		0xB7
+#define REG0_DATA 0xB1
+#define REG1_DATA 0xB3
+#define REG2_DATA 0xB5
+#define REG3_DATA 0xB7
 
 /* bit definitions of Measurement Enable Register 2 (0x51) */
-#define MEAS2_VRTC		(1 << 0)
+#define MEAS2_VRTC (1 << 0)
 
 /* bit definitions of RTC Register 1 (0xA0) */
-#define ALARM_EN		(1 << 3)
-#define ALARM_WAKEUP		(1 << 4)
-#define ALARM			(1 << 5)
-#define RTC1_USE_XO		(1 << 7)
+#define ALARM_EN (1 << 3)
+#define ALARM_WAKEUP (1 << 4)
+#define ALARM (1 << 5)
+#define RTC1_USE_XO (1 << 7)
 
-#define VRTC_CALIB_INTERVAL	(HZ * 60 * 10)		/* 10 minutes */
+#define VRTC_CALIB_INTERVAL (HZ * 60 * 10) /* 10 minutes */
 
 static irqreturn_t rtc_update_handler(int irq, void *data)
 {
@@ -83,7 +84,7 @@ static int pm860x_rtc_alarm_irq_enable(struct device *dev, unsigned int enabled)
  * and the current time.
  */
 static void rtc_next_alarm_time(struct rtc_time *next, struct rtc_time *now,
-				struct rtc_time *alrm)
+								struct rtc_time *alrm)
 {
 	unsigned long next_time;
 	unsigned long now_time;
@@ -98,7 +99,8 @@ static void rtc_next_alarm_time(struct rtc_time *next, struct rtc_time *now,
 	rtc_tm_to_time(now, &now_time);
 	rtc_tm_to_time(next, &next_time);
 
-	if (next_time < now_time) {
+	if (next_time < now_time)
+	{
 		/* Advance one day */
 		next_time += 60 * 60 * 24;
 		rtc_time_to_tm(next_time, next);
@@ -113,15 +115,17 @@ static int pm860x_rtc_read_time(struct device *dev, struct rtc_time *tm)
 
 	pm860x_page_bulk_read(info->i2c, REG0_ADDR, 8, buf);
 	dev_dbg(info->dev, "%x-%x-%x-%x-%x-%x-%x-%x\n", buf[0], buf[1],
-		buf[2], buf[3], buf[4], buf[5], buf[6], buf[7]);
-	base = (buf[1] << 24) | (buf[3] << 16) | (buf[5] << 8) | buf[7];
+			buf[2], buf[3], buf[4], buf[5], buf[6], buf[7]);
+	base = ((unsigned long)buf[1] << 24) | (buf[3] << 16) |
+		   (buf[5] << 8) | buf[7];
 
 	/* load 32-bit read-only counter */
 	pm860x_bulk_read(info->i2c, PM8607_RTC_COUNTER1, 4, buf);
-	data = (buf[3] << 24) | (buf[2] << 16) | (buf[1] << 8) | buf[0];
+	data = ((unsigned long)buf[3] << 24) | (buf[2] << 16) |
+		   (buf[1] << 8) | buf[0];
 	ticks = base + data;
 	dev_dbg(info->dev, "get base:0x%lx, RO count:0x%lx, ticks:0x%lx\n",
-		base, data, ticks);
+			base, data, ticks);
 
 	rtc_time_to_tm(ticks, tm);
 
@@ -134,20 +138,22 @@ static int pm860x_rtc_set_time(struct device *dev, struct rtc_time *tm)
 	unsigned char buf[4];
 	unsigned long ticks, base, data;
 
-	if ((tm->tm_year < 70) || (tm->tm_year > 138)) {
+	if ((tm->tm_year < 70) || (tm->tm_year > 138))
+	{
 		dev_dbg(info->dev, "Set time %d out of range. "
-			"Please set time between 1970 to 2038.\n",
-			1900 + tm->tm_year);
+						   "Please set time between 1970 to 2038.\n",
+				1900 + tm->tm_year);
 		return -EINVAL;
 	}
 	rtc_tm_to_time(tm, &ticks);
 
 	/* load 32-bit read-only counter */
 	pm860x_bulk_read(info->i2c, PM8607_RTC_COUNTER1, 4, buf);
-	data = (buf[3] << 24) | (buf[2] << 16) | (buf[1] << 8) | buf[0];
+	data = ((unsigned long)buf[3] << 24) | (buf[2] << 16) |
+		   (buf[1] << 8) | buf[0];
 	base = ticks - data;
 	dev_dbg(info->dev, "set base:0x%lx, RO count:0x%lx, ticks:0x%lx\n",
-		base, data, ticks);
+			base, data, ticks);
 
 	pm860x_page_reg_write(info->i2c, REG0_DATA, (base >> 24) & 0xFF);
 	pm860x_page_reg_write(info->i2c, REG1_DATA, (base >> 16) & 0xFF);
@@ -168,14 +174,16 @@ static int pm860x_rtc_read_alarm(struct device *dev, struct rtc_wkalrm *alrm)
 
 	pm860x_page_bulk_read(info->i2c, REG0_ADDR, 8, buf);
 	dev_dbg(info->dev, "%x-%x-%x-%x-%x-%x-%x-%x\n", buf[0], buf[1],
-		buf[2], buf[3], buf[4], buf[5], buf[6], buf[7]);
-	base = (buf[1] << 24) | (buf[3] << 16) | (buf[5] << 8) | buf[7];
+			buf[2], buf[3], buf[4], buf[5], buf[6], buf[7]);
+	base = ((unsigned long)buf[1] << 24) | (buf[3] << 16) |
+		   (buf[5] << 8) | buf[7];
 
 	pm860x_bulk_read(info->i2c, PM8607_RTC_EXPIRE1, 4, buf);
-	data = (buf[3] << 24) | (buf[2] << 16) | (buf[1] << 8) | buf[0];
+	data = ((unsigned long)buf[3] << 24) | (buf[2] << 16) |
+		   (buf[1] << 8) | buf[0];
 	ticks = base + data;
 	dev_dbg(info->dev, "get base:0x%lx, RO count:0x%lx, ticks:0x%lx\n",
-		base, data, ticks);
+			base, data, ticks);
 
 	rtc_time_to_tm(ticks, &alrm->time);
 	ret = pm860x_reg_read(info->i2c, PM8607_RTC1);
@@ -196,15 +204,17 @@ static int pm860x_rtc_set_alarm(struct device *dev, struct rtc_wkalrm *alrm)
 
 	pm860x_page_bulk_read(info->i2c, REG0_ADDR, 8, buf);
 	dev_dbg(info->dev, "%x-%x-%x-%x-%x-%x-%x-%x\n", buf[0], buf[1],
-		buf[2], buf[3], buf[4], buf[5], buf[6], buf[7]);
-	base = (buf[1] << 24) | (buf[3] << 16) | (buf[5] << 8) | buf[7];
+			buf[2], buf[3], buf[4], buf[5], buf[6], buf[7]);
+	base = ((unsigned long)buf[1] << 24) | (buf[3] << 16) |
+		   (buf[5] << 8) | buf[7];
 
 	/* load 32-bit read-only counter */
 	pm860x_bulk_read(info->i2c, PM8607_RTC_COUNTER1, 4, buf);
-	data = (buf[3] << 24) | (buf[2] << 16) | (buf[1] << 8) | buf[0];
+	data = ((unsigned long)buf[3] << 24) | (buf[2] << 16) |
+		   (buf[1] << 8) | buf[0];
 	ticks = base + data;
 	dev_dbg(info->dev, "get base:0x%lx, RO count:0x%lx, ticks:0x%lx\n",
-		base, data, ticks);
+			base, data, ticks);
 
 	rtc_time_to_tm(ticks, &now_tm);
 	rtc_next_alarm_time(&alarm_tm, &now_tm, &alrm->time);
@@ -217,22 +227,25 @@ static int pm860x_rtc_set_alarm(struct device *dev, struct rtc_wkalrm *alrm)
 	buf[2] = (data >> 16) & 0xff;
 	buf[3] = (data >> 24) & 0xff;
 	pm860x_bulk_write(info->i2c, PM8607_RTC_EXPIRE1, 4, buf);
-	if (alrm->enabled) {
+	if (alrm->enabled)
+	{
 		mask = ALARM | ALARM_WAKEUP | ALARM_EN;
 		pm860x_set_bits(info->i2c, PM8607_RTC1, mask, mask);
-	} else {
+	}
+	else
+	{
 		mask = ALARM | ALARM_WAKEUP | ALARM_EN;
 		pm860x_set_bits(info->i2c, PM8607_RTC1, mask,
-				ALARM | ALARM_WAKEUP);
+						ALARM | ALARM_WAKEUP);
 	}
 	return 0;
 }
 
 static const struct rtc_class_ops pm860x_rtc_ops = {
-	.read_time	= pm860x_rtc_read_time,
-	.set_time	= pm860x_rtc_set_time,
-	.read_alarm	= pm860x_rtc_read_alarm,
-	.set_alarm	= pm860x_rtc_set_alarm,
+	.read_time = pm860x_rtc_read_time,
+	.set_time = pm860x_rtc_set_time,
+	.read_alarm = pm860x_rtc_read_alarm,
+	.set_alarm = pm860x_rtc_set_alarm,
 	.alarm_irq_enable = pm860x_rtc_alarm_irq_enable,
 };
 
@@ -240,16 +253,17 @@ static const struct rtc_class_ops pm860x_rtc_ops = {
 static void calibrate_vrtc_work(struct work_struct *work)
 {
 	struct pm860x_rtc_info *info = container_of(work,
-		struct pm860x_rtc_info, calib_work.work);
+												struct pm860x_rtc_info, calib_work.work);
 	unsigned char buf[2];
 	unsigned int sum, data, mean, vrtc_set;
 	int i;
 
-	for (i = 0, sum = 0; i < 16; i++) {
+	for (i = 0, sum = 0; i < 16; i++)
+	{
 		msleep(100);
 		pm860x_bulk_read(info->i2c, REG_VRTC_MEAS1, 2, buf);
 		data = (buf[0] << 4) | buf[1];
-		data = (data * 5400) >> 12;	/* convert to mv */
+		data = (data * 5400) >> 12; /* convert to mv */
 		sum += data;
 	}
 	mean = sum >> 4;
@@ -258,19 +272,23 @@ static void calibrate_vrtc_work(struct work_struct *work)
 
 	sum = pm860x_reg_read(info->i2c, PM8607_RTC_MISC1);
 	data = sum & 0x3;
-	if ((mean + 200) < vrtc_set) {
+	if ((mean + 200) < vrtc_set)
+	{
 		/* try higher voltage */
 		if (++data == 4)
 			goto out;
 		data = (sum & 0xf8) | (data & 0x3);
 		pm860x_reg_write(info->i2c, PM8607_RTC_MISC1, data);
-	} else if ((mean - 200) > vrtc_set) {
+	}
+	else if ((mean - 200) > vrtc_set)
+	{
 		/* try lower voltage */
 		if (data-- == 0)
 			goto out;
 		data = (sum & 0xf8) | (data & 0x3);
 		pm860x_reg_write(info->i2c, PM8607_RTC_MISC1, data);
-	} else
+	}
+	else
 		goto out;
 	dev_dbg(info->dev, "set 0x%x to RTC_MISC1\n", data);
 	/* trigger next calibration since VRTC is updated */
@@ -301,7 +319,8 @@ static int __devinit pm860x_rtc_probe(struct platform_device *pdev)
 	if (!info)
 		return -ENOMEM;
 	info->irq = platform_get_irq(pdev, 0);
-	if (info->irq < 0) {
+	if (info->irq < 0)
+	{
 		dev_err(&pdev->dev, "No IRQ resource!\n");
 		ret = -EINVAL;
 		goto out;
@@ -313,10 +332,11 @@ static int __devinit pm860x_rtc_probe(struct platform_device *pdev)
 	dev_set_drvdata(&pdev->dev, info);
 
 	ret = request_threaded_irq(info->irq, NULL, rtc_update_handler,
-				   IRQF_ONESHOT, "rtc", info);
-	if (ret < 0) {
+							   IRQF_ONESHOT, "rtc", info);
+	if (ret < 0)
+	{
 		dev_err(chip->dev, "Failed to request IRQ: #%d: %d\n",
-			info->irq, ret);
+				info->irq, ret);
 		goto out;
 	}
 
@@ -327,11 +347,13 @@ static int __devinit pm860x_rtc_probe(struct platform_device *pdev)
 	pm860x_page_reg_write(info->i2c, REG3_ADDR, REG3_DATA);
 
 	ret = pm860x_rtc_read_time(&pdev->dev, &tm);
-	if (ret < 0) {
+	if (ret < 0)
+	{
 		dev_err(&pdev->dev, "Failed to read initial time.\n");
 		goto out_rtc;
 	}
-	if ((tm.tm_year < 70) || (tm.tm_year > 138)) {
+	if ((tm.tm_year < 70) || (tm.tm_year > 138))
+	{
 		tm.tm_year = 70;
 		tm.tm_mon = 0;
 		tm.tm_mday = 1;
@@ -339,21 +361,24 @@ static int __devinit pm860x_rtc_probe(struct platform_device *pdev)
 		tm.tm_min = 0;
 		tm.tm_sec = 0;
 		ret = pm860x_rtc_set_time(&pdev->dev, &tm);
-		if (ret < 0) {
+		if (ret < 0)
+		{
 			dev_err(&pdev->dev, "Failed to set initial time.\n");
 			goto out_rtc;
 		}
 	}
 	rtc_tm_to_time(&tm, &ticks);
-	if (pdata && pdata->sync) {
+	if (pdata && pdata->sync)
+	{
 		pdata->sync(ticks);
 		info->sync = pdata->sync;
 	}
 
 	info->rtc_dev = rtc_device_register("88pm860x-rtc", &pdev->dev,
-					    &pm860x_rtc_ops, THIS_MODULE);
+										&pm860x_rtc_ops, THIS_MODULE);
 	ret = PTR_ERR(info->rtc_dev);
-	if (IS_ERR(info->rtc_dev)) {
+	if (IS_ERR(info->rtc_dev))
+	{
 		dev_err(&pdev->dev, "Failed to register RTC device: %d\n", ret);
 		goto out_rtc;
 	}
@@ -375,7 +400,7 @@ static int __devinit pm860x_rtc_probe(struct platform_device *pdev)
 	/* calibrate VRTC */
 	INIT_DELAYED_WORK(&info->calib_work, calibrate_vrtc_work);
 	schedule_delayed_work(&info->calib_work, VRTC_CALIB_INTERVAL);
-#endif	/* VRTC_CALIBRATION */
+#endif /* VRTC_CALIBRATION */
 
 	device_init_wakeup(&pdev->dev, 1);
 
@@ -395,7 +420,7 @@ static int __devexit pm860x_rtc_remove(struct platform_device *pdev)
 	flush_scheduled_work();
 	/* disable measurement */
 	pm860x_set_bits(info->i2c, PM8607_MEAS_EN2, MEAS2_VRTC, 0);
-#endif	/* VRTC_CALIBRATION */
+#endif /* VRTC_CALIBRATION */
 
 	platform_set_drvdata(pdev, NULL);
 	rtc_device_unregister(info->rtc_dev);
@@ -428,13 +453,13 @@ static int pm860x_rtc_resume(struct device *dev)
 static SIMPLE_DEV_PM_OPS(pm860x_rtc_pm_ops, pm860x_rtc_suspend, pm860x_rtc_resume);
 
 static struct platform_driver pm860x_rtc_driver = {
-	.driver		= {
-		.name	= "88pm860x-rtc",
-		.owner	= THIS_MODULE,
-		.pm	= &pm860x_rtc_pm_ops,
+	.driver = {
+		.name = "88pm860x-rtc",
+		.owner = THIS_MODULE,
+		.pm = &pm860x_rtc_pm_ops,
 	},
-	.probe		= pm860x_rtc_probe,
-	.remove		= __devexit_p(pm860x_rtc_remove),
+	.probe = pm860x_rtc_probe,
+	.remove = __devexit_p(pm860x_rtc_remove),
 };
 
 module_platform_driver(pm860x_rtc_driver);
