+++ drivers/rtc/rtc-ds1672.c
@@ -17,11 +17,11 @@
 
 /* Registers */
 
-#define DS1672_REG_CNT_BASE	0
-#define DS1672_REG_CONTROL	4
-#define DS1672_REG_TRICKLE	5
+#define DS1672_REG_CNT_BASE 0
+#define DS1672_REG_CONTROL 4
+#define DS1672_REG_TRICKLE 5
 
-#define DS1672_REG_CONTROL_EOSC	0x80
+#define DS1672_REG_CONTROL_EOSC 0x80
 
 static struct i2c_driver ds1672_driver;
 
@@ -37,28 +37,30 @@ static int ds1672_get_datetime(struct i2c_client *client, struct rtc_time *tm)
 	unsigned char buf[4];
 
 	struct i2c_msg msgs[] = {
-		{client->addr, 0, 1, &addr},	/* setup read ptr */
-		{client->addr, I2C_M_RD, 4, buf},	/* read date */
+		{client->addr, 0, 1, &addr},	  /* setup read ptr */
+		{client->addr, I2C_M_RD, 4, buf}, /* read date */
 	};
 
 	/* read date registers */
-	if ((i2c_transfer(client->adapter, &msgs[0], 2)) != 2) {
+	if ((i2c_transfer(client->adapter, &msgs[0], 2)) != 2)
+	{
 		dev_err(&client->dev, "%s: read error\n", __func__);
 		return -EIO;
 	}
 
 	dev_dbg(&client->dev,
-		"%s: raw read data - counters=%02x,%02x,%02x,%02x\n",
-		__func__, buf[0], buf[1], buf[2], buf[3]);
+			"%s: raw read data - counters=%02x,%02x,%02x,%02x\n",
+			__func__, buf[0], buf[1], buf[2], buf[3]);
 
-	time = (buf[3] << 24) | (buf[2] << 16) | (buf[1] << 8) | buf[0];
+	time = ((unsigned long)buf[3] << 24) | (buf[2] << 16) |
+		   (buf[1] << 8) | buf[0];
 
 	rtc_time_to_tm(time, tm);
 
 	dev_dbg(&client->dev, "%s: tm is secs=%d, mins=%d, hours=%d, "
-		"mday=%d, mon=%d, year=%d, wday=%d\n",
-		__func__, tm->tm_sec, tm->tm_min, tm->tm_hour,
-		tm->tm_mday, tm->tm_mon, tm->tm_year, tm->tm_wday);
+						  "mday=%d, mon=%d, year=%d, wday=%d\n",
+			__func__, tm->tm_sec, tm->tm_min, tm->tm_hour,
+			tm->tm_mday, tm->tm_mon, tm->tm_year, tm->tm_wday);
 
 	return 0;
 }
@@ -73,10 +75,11 @@ static int ds1672_set_mmss(struct i2c_client *client, unsigned long secs)
 	buf[2] = (secs & 0x0000FF00) >> 8;
 	buf[3] = (secs & 0x00FF0000) >> 16;
 	buf[4] = (secs & 0xFF000000) >> 24;
-	buf[5] = 0;		/* set control reg to enable counting */
+	buf[5] = 0; /* set control reg to enable counting */
 
 	xfer = i2c_master_send(client, buf, 6);
-	if (xfer != 6) {
+	if (xfer != 6)
+	{
 		dev_err(&client->dev, "%s: send: %d\n", __func__, xfer);
 		return -EIO;
 	}
@@ -99,12 +102,13 @@ static int ds1672_get_control(struct i2c_client *client, u8 *status)
 	unsigned char addr = DS1672_REG_CONTROL;
 
 	struct i2c_msg msgs[] = {
-		{client->addr, 0, 1, &addr},	/* setup read ptr */
-		{client->addr, I2C_M_RD, 1, status},	/* read control */
+		{client->addr, 0, 1, &addr},		 /* setup read ptr */
+		{client->addr, I2C_M_RD, 1, status}, /* read control */
 	};
 
 	/* read control register */
-	if ((i2c_transfer(client->adapter, &msgs[0], 2)) != 2) {
+	if ((i2c_transfer(client->adapter, &msgs[0], 2)) != 2)
+	{
 		dev_err(&client->dev, "%s: read error\n", __func__);
 		return -EIO;
 	}
@@ -114,7 +118,7 @@ static int ds1672_get_control(struct i2c_client *client, u8 *status)
 
 /* following are the sysfs callback functions */
 static ssize_t show_control(struct device *dev, struct device_attribute *attr,
-			    char *buf)
+							char *buf)
 {
 	struct i2c_client *client = to_i2c_client(dev);
 	u8 control;
@@ -124,8 +128,7 @@ static ssize_t show_control(struct device *dev, struct device_attribute *attr,
 	if (err)
 		return err;
 
-	return sprintf(buf, "%s\n", (control & DS1672_REG_CONTROL_EOSC)
-		       ? "disabled" : "enabled");
+	return sprintf(buf, "%s\n", (control & DS1672_REG_CONTROL_EOSC) ? "disabled" : "enabled");
 }
 
 static DEVICE_ATTR(control, S_IRUGO, show_control, NULL);
@@ -146,7 +149,7 @@ static int ds1672_remove(struct i2c_client *client)
 }
 
 static int ds1672_probe(struct i2c_client *client,
-			const struct i2c_device_id *id)
+						const struct i2c_device_id *id)
 {
 	int err = 0;
 	u8 control;
@@ -160,7 +163,7 @@ static int ds1672_probe(struct i2c_client *client,
 	dev_info(&client->dev, "chip found, driver version " DRV_VERSION "\n");
 
 	rtc = rtc_device_register(ds1672_driver.driver.name, &client->dev,
-				  &ds1672_rtc_ops, THIS_MODULE);
+							  &ds1672_rtc_ops, THIS_MODULE);
 
 	if (IS_ERR(rtc))
 		return PTR_ERR(rtc);
@@ -174,7 +177,7 @@ static int ds1672_probe(struct i2c_client *client,
 
 	if (control & DS1672_REG_CONTROL_EOSC)
 		dev_warn(&client->dev, "Oscillator not enabled. "
-			 "Set time to enable.\n");
+							   "Set time to enable.\n");
 
 	/* Register sysfs hooks */
 	err = device_create_file(&client->dev, &dev_attr_control);
@@ -183,20 +186,19 @@ static int ds1672_probe(struct i2c_client *client,
 
 	return 0;
 
- exit_devreg:
+exit_devreg:
 	rtc_device_unregister(rtc);
 	return err;
 }
 
 static struct i2c_device_id ds1672_id[] = {
-	{ "ds1672", 0 },
-	{ }
-};
+	{"ds1672", 0},
+	{}};
 
 static struct i2c_driver ds1672_driver = {
 	.driver = {
-		   .name = "rtc-ds1672",
-		   },
+		.name = "rtc-ds1672",
+	},
 	.probe = &ds1672_probe,
 	.remove = &ds1672_remove,
 	.id_table = ds1672_id,
